% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activos_en.R
\name{activos_en}
\alias{activos_en}
\title{Cuenta los sindicatos en constitución o activos en cada periodo}
\usage{
activos_en(ano = 2021, mes = 6)
}
\description{
Cuenta los sindicatos en constitución o activos en cada periodo
}
\examples{

activos_en(ano=1990,mes=1)

}
