% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasa_sindicalizacion.R
\name{afiliados_tipo}
\alias{afiliados_tipo}
\title{Afiliados según tipo de sindicato}
\usage{
afiliados_tipo()
}
\value{
Una tibble con afiliados a sindicatos según tipo de sindicatos
}
\description{
Imprime el número de afiliados/as a sindicatos según tipo de sindicato desde 1990
hasta 2020.
}
\examples{
afiliados_tipo()

}
