% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscar_sindicatos.R
\name{buscar_sindicatos}
\alias{buscar_sindicatos}
\title{Buscar empresas o sindicatos por palabras claves}
\usage{
buscar_sindicatos(empresa_buscada = "", sindicato_buscado = "")
}
\description{
Buscar empresas o sindicatos por palabras claves
}
\examples{
buscar_sindicatos(empresa_buscada="walmart",sindicato_buscado="lider")
buscar_sindicatos(empresa_buscada="codelco",sindicato_buscado="")
buscar_sindicatos(empresa_buscada="",sindicato_buscado="casa particular")
}
