% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrupar_empresa.R
\name{empresas_sindicatos_activos}
\alias{empresas_sindicatos_activos}
\title{Agrupa por empresa e indica N de sindicatos activos y afiliados/as}
\usage{
empresas_sindicatos_activos()
}
\value{
Una tibble en la que cada fila es una empresa
}
\description{
Además reporta para cada empresa el tamaño promedio de los sindicatos.
}
\examples{
empresas_sindicatos_activos()

}
