% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sii.R
\docType{data}
\name{sii}
\alias{sii}
\title{SII}
\format{

}
\source{
\url{https://www.sii.cl/sobre_el_sii/nominapersonasjuridicas.html}
}
\usage{
sii
}
\description{
Datos del Servicio de Impuestos Internos, consolidado de años 2018 y 2019.
Cada fila es una empresa, no existen duplicados.
El dato para cada empresa es el más actual del SII.
}
\details{
Un contribuyente es clasificado como empresa,
cuando cumpla con uno o más de los siguientes atributos:
Es identificado como contribuyente de primera vategoría,
presenta declaración jurada 1887 y/o es declarante vigente de IVA
}
\examples{

}
\keyword{datasets}
