% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sii_agregar.R
\name{sii_agregar}
\alias{sii_agregar}
\title{Agrega información del Servicio de impuestos internos}
\usage{
sii_agregar()
}
\description{
Se filtra la data a los sindicatos activos y de empresa o establecimiento.
Luego se agregan variables del Servicio de Impuestos Internos relacionadas a
ventas, número de trabajadores y actividad económica.
}
\examples{

datos <- sii_agregar()
table(datos$rubro_sii,useNA="ifany")
table(datos$ventas_sii,useNA="ifany")
nrow(datos[datos$tasa_sindicalizacion<=1&!is.na(datos$tasa_sindicalizacion),])
summary(datos[datos$tasa_sindicalizacion<=1,]$tasa_sindicalizacion)
}
