% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sindicatos_activos.R
\name{sindicatos_activos}
\alias{sindicatos_activos}
\title{Filtra sindicatos activos}
\usage{
sindicatos_activos()
}
\value{
Una tibble con listado de sindicatos activos
}
\description{
Filtra sindicatos activos
}
\examples{
sindicatos_activos()

}
