% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sindicatos_activos.R
\name{sindicatos_activos_empresas}
\alias{sindicatos_activos_empresas}
\title{Filtrar sindicatos según agrupaciones relevantes}
\usage{
sindicatos_activos_empresas()
}
\description{
Filtrar sindicatos según agrupaciones relevantes
}
\examples{
sindicatos_activos_empresas()

}
