% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sindicatos_empresas.R
\name{sindicatos_empresas}
\alias{sindicatos_empresas}
\title{Filtrar sindicatos según agrupaciones relevantes}
\usage{
sindicatos_empresas()
}
\description{
Filtrar sindicatos según agrupaciones relevantes
}
\examples{
sindicatos_activos()
sindicatos_empresas()

}
