% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sindicatos_independientes.R
\name{sindicatos_independientes}
\alias{sindicatos_independientes}
\title{Filtra sindicatos según independientes}
\usage{
sindicatos_independientes()
}
\value{
Filtra sindicatos dejando solo a los de empresas independientes
}
\description{
Filtra sindicatos según independientes
}
\examples{
sindicatos_activos()
sindicatos_empresas()

}
