% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sindicatos_interempresas.R
\name{sindicatos_interempresas}
\alias{sindicatos_interempresas}
\title{Filtrar sindicatos según agrupaciones relevantes}
\usage{
sindicatos_interempresas()
}
\description{
Filtrar sindicatos según agrupaciones relevantes
}
\examples{
sindicatos_activos()
sindicatos_empresas()

}
