% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasa_sindicalizacion.R
\name{tasa_sindicalizacion}
\alias{tasa_sindicalizacion}
\title{Tasa de afiliación sindical según tres metodologías de la Dirección del Trabajo}
\usage{
tasa_sindicalizacion(metodo = "res")
}
\arguments{
\item{metodo}{puede "res2 o "dt"}
}
\description{
dt = imprime las tasas oficiales de la DT publicada en compendios
res = imprime las tasas según metodología DT pero con datos de sindicatoscl
}
\details{
metodologia1 = Población Afiliada / Total de ocupados*100
metodologia2 = Población Afiliada / Total de ocupados (excluye asalariados s.público)*100
metodologia3 = Población Afiliada (excluye afiliados independientes) / Total de ocupados
(excluye asalariados s.público)*100
}
\examples{
tasa_sindicalizacion()

}
