% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activos_en.R
\name{activos_en}
\alias{activos_en}
\title{Cuenta los sindicatos en constituci칩n o activos en cada periodo (a침o o mes)}
\usage{
activos_en(ano = 2021, mes = 6)
}
\description{
Cuenta los sindicatos en constituci칩n o activos en cada periodo (a침o o mes)
}
\examples{
constitucion_anual()
activos_historico(desde=2020,hasta=2021)
activos_en(ano=1990,mes=1)

}
