% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activos_en.R
\name{activos_en}
\alias{activos_en}
\title{Cuenta los sindicatos en constitución o activos en cada periodo}
\usage{
activos_en(ano = 2021, mes = 11)
}
\arguments{
\item{ano}{Año en el que se buscan sindicatos activos}

\item{mes}{Mes en el que se buscan sindicatos activos. Por defecto noviembre 2021}
}
\value{
una tibble con los sindicatos activos en el periodo consultado
}
\description{
Cuenta los sindicatos en constitución o activos en cada periodo
}
\examples{

activos_en(ano=1990,mes=1)
activos_en(ano=1990)
}
