% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activos_historico.R
\name{activos_historico}
\alias{activos_historico}
\title{Cuenta los sindicatos en constituci칩n o activos en cada periodo (a침o o mes)}
\usage{
activos_historico(desde = 2020, hasta = 2021)
}
\description{
Cuenta los sindicatos en constituci칩n o activos en cada periodo (a침o o mes)
}
\examples{
activos_historico(desde=2020,hasta=2021)


}
