% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buscar_sindicatos.R
\name{buscar_sindicatos}
\alias{buscar_sindicatos}
\title{Buscar empresas o sindicatos por palabras claves}
\usage{
buscar_sindicatos(empresa_buscada = "", sindicato_buscado = "")
}
\arguments{
\item{empresa_buscada}{cadena de texto con palabra clave}

\item{sindicato_buscado}{cadena de texto con palabra clave}
}
\value{
Una tibble con los sindicatos filtrados
}
\description{
Permite buscar sindicatos, ya sea activos o en receso, en empresas según
el nombre del sindicato y/o de la empresa
}
\examples{
buscar_sindicatos(empresa_buscada="walmart",sindicato_buscado="lider")
buscar_sindicatos(empresa_buscada="codelco",sindicato_buscado="")
buscar_sindicatos(empresa_buscada="",sindicato_buscado="casa particular")
}
