% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constitucion_anual.R
\name{constitucion_anual}
\alias{constitucion_anual}
\title{Cuenta los sindicatos en constituci칩n o activos en cada periodo (a침o o mes)}
\usage{
constitucion_anual()
}
\description{
Cuenta los sindicatos en constituci칩n o activos en cada periodo (a침o o mes)
}
\examples{
constitucion_anual()

}
