% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrupar_empresa.R
\name{empresas_sindicatos_activos}
\alias{empresas_sindicatos_activos}
\title{Agrupa por empresa e indica N de sindicatos activos y afiliados/as}
\usage{
empresas_sindicatos_activos()
}
\description{
Agrupa por empresa e indica N de sindicatos activos y afiliados/as
}
\examples{
empresas_sindicatos_activos()

}
