% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sindicatos_activos.R
\name{sindicatos_activos}
\alias{sindicatos_activos}
\title{Filtrar sindicatos según agrupaciones relevantes}
\usage{
sindicatos_activos()
}
\description{
Filtrar sindicatos según agrupaciones relevantes
}
\examples{
sindicatos_activos()
sindicatos_empresas()

}
