% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sindicatos_activos.R
\name{sindicatos_activos_empresas}
\alias{sindicatos_activos_empresas}
\title{Filtra sindicatos activos y de empresa o establecimiento}
\usage{
sindicatos_activos_empresas()
}
\value{
Una tibble con listado de sindicatos activos y de empresa
}
\description{
Filtra sindicatos activos y de empresa o establecimiento
}
\examples{
sindicatos_activos_empresas()

}
