% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sindicatos_empresas.R
\name{sindicatos_empresas}
\alias{sindicatos_empresas}
\title{Filtrar sindicatos de empresa o establecimiento.}
\usage{
sindicatos_empresas()
}
\value{
Una tibble
}
\description{
No distingue entre receso y activos
}
\examples{
sindicatos_activos()
sindicatos_empresas()

}
