% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sindicatos_independientes.R
\name{sindicatos_independientes}
\alias{sindicatos_independientes}
\title{Filtrar sindicatos según agrupaciones relevantes}
\usage{
sindicatos_independientes()
}
\description{
Filtrar sindicatos según agrupaciones relevantes
}
\examples{
sindicatos_activos()
sindicatos_empresas()

}
