% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sindicatos_interempresas.R
\name{sindicatos_interempresas}
\alias{sindicatos_interempresas}
\title{Filtra sindicatos dejando solo a los interempresa}
\usage{
sindicatos_interempresas()
}
\value{
Una tibble
}
\description{
Filtra sindicatos dejando solo a los interempresa
}
\examples{
sindicatos_activos()
sindicatos_empresas()

}
